+++ packages/TIFFImageryProvider/src/plotty/typing.ts
@@ -15,6 +15,15 @@ export type RenderColorType = 'continuous' | 'discrete';
 
 export type ColorScaleNames = keyof typeof colorscales;
 
+export interface RGBOptions {
+  bands: {
+    r?: { band: number; min?: number; max?: number; };
+    g?: { band: number; min?: number; max?: number; };
+    b?: { band: number; min?: number; max?: number; };
+  };
+  colorMapping?: Record<string, string>;
+}
+
 export type PlotOptions = {
   /**
    * The canvas to render to.
@@ -35,6 +44,21 @@ export type PlotOptions = {
    * The height of the input raster.
    */
   height?: number;
+
+  /**
+   * The width of the output tile.
+   */
+  tileWidth?: number;
+
+  /**
+   * The height of the output tile.
+   */
+  tileHeight?: number;
+
+  /**
+   * The buffer of the input tile.
+   */
+  buffer?: number;
 
   /**
    * A list of named datasets. Each must have 'id', 'data', 'width' and 'height'.
@@ -60,4 +84,6 @@ export type PlotOptions = {
    * Plotty can also function with pure javascript but it is much slower then using WebGL rendering.
    */
   useWebGL?: boolean;
+
+  rgbOptions?: RGBOptions;
 } & SingleBandRenderOptions