+++ packages/TIFFImageryProvider/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.1](https://github.com/hongfaqiu/tiff-imagery-provider/compare/v2.15.0...v2.15.1) (2024-11-15)
+
+
+### Bug Fixes
+
+* process NAN value ([cb91df3](https://github.com/hongfaqiu/tiff-imagery-provider/commit/cb91df38ddfd67911a8e73c4c6bdf0eb054cc5b9))
+
+
+
+
+
+# [2.15.0](https://github.com/hongfaqiu/tiff-imagery-provider/compare/v2.14.2...v2.15.0) (2024-11-15)
+
+
+### Features
+
+* **plot:** use plot class for RGB rendering ([#41](https://github.com/hongfaqiu/tiff-imagery-provider/issues/41)) ([c133785](https://github.com/hongfaqiu/tiff-imagery-provider/commit/c1337856902b9dec5f79797c8acb4e2295a66362))
+
+
+
+
+
+
+## [2.14.2](https://github.com/hongfaqiu/tiff-imagery-provider/compare/v2.14.1...v2.14.2) (2024-10-18)
+
+
+### Bug Fixes
+
+* modify getValue function to properly handle buffer boundaries ([267cc0a](https://github.com/hongfaqiu/tiff-imagery-provider/commit/267cc0a1913d2c2dd5ec72e6c5f19587ca4ee465))
+
+
+
+
+
+## [2.14.1](https://github.com/hongfaqiu/tiff-imagery-provider/compare/v2.14.0...v2.14.1) (2024-10-18)
+
+
+### Bug Fixes
+
+* copyNewSize func ([df4fde7](https://github.com/hongfaqiu/tiff-imagery-provider/commit/df4fde7d689523b8720822a6846f0719ec6863cc))
+
+
+
+
+
+# [2.14.0](https://github.com/hongfaqiu/tiff-imagery-provider/compare/v2.13.3...v2.14.0) (2024-10-18)
+
+
+### Features
+
+* add useImageCountAsMaximumLevel parameter ([8ffc2c2](https://github.com/hongfaqiu/tiff-imagery-provider/commit/8ffc2c2d0a9c531b19a345011b61c726909be1ea))
+* gpu based resample ([0ca44b2](https://github.com/hongfaqiu/tiff-imagery-provider/commit/0ca44b26e355dc1c61fd7210250dcef73af62990))
+
+
+
+
+
 ## [2.13.3](https://github.com/hongfaqiu/tiff-imagery-provider/compare/v2.13.2...v2.13.3) (2024-10-11)